feat: change folder structure from pages to app #34
