+++ .prettierignore
@@ -1,4 +1,4 @@
-# 빌드·배포 산출물
+# 빌드·배포 산출물Add commentMore actions
 .next
 out
 coverage
